backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
python-dotenv==1.0.0
Pillow>=10.3.0
psutil==5.9.7
Werkzeug==3.0.1
requests>=2.31.0
